requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coverage==7.3.1
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.5
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
exceptiongroup==1.1.3
Faker==19.6.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
oauthlib==3.2.2
packaging==23.1
pluggy==1.3.0
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.2
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
